html/interaction/focus/sequential-focus-navigation-and-the-tabindex-attribute/focus-tabindex-order-shadow-varying-tabindex.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>HTML Test: focus - the sequential focus navigation order with shadow dom with varying tabindex values</title>
<link rel="help" href="https://html.spec.whatwg.org/multipage/interaction.html#sequential-focus-navigation">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="resources/shadow-utils.js"></script>
<body>
<script>
// Structure:
// <div #aboveHost tabindex=3>
// <div #host tabindex=3>
//    #shadowRoot
//      <div #aboveSlot tabindex=2>
//      <slot #slotAbove tabindex=1>
//        (slotted) <div #slottedAbove tabindex=4>
//      <slot #slotBelow tabindex=1>
//        (slotted) <div #slottedBelow tabindex=4>
//      <div #belowSlot tabindex=2>
// <div #belowHost tabindex=3>

async_test(async (t) => {
  let elementsInFlatTreeOrder;
  let [aboveHost, host, aboveSlot, slotAbove, slottedAbove, slotBelow, slottedBelow, belowSlot, belowHost] =
    elementsInFlatTreeOrder = prepareDOM(document.body, false);
  setTabIndex([slotAbove, slotBelow], 1);
  setTabIndex([aboveSlot, belowSlot], 2);
  setTabIndex([aboveHost, host, belowHost], 3);
  setTabIndex([slottedAbove, slottedBelow], 4);
  resetFocus();
  const expectedOrder = [aboveHost, host, slottedAbove, slottedBelow, aboveSlot, belowSlot, belowHost];
  for (let el of expectedOrder) {
    await navigateFocusForward();
    let focused = host.shadowRoot.activeElement ? host.shadowRoot.activeElement : document.activeElement;
    t.step(() => { assert_equals(focused, el) });
  }
  t.done();
}, "Order with various tabindex values");
</script>
</body>
